Don't trim the message
This is useful when piping something in via `echo` but if we're piping
in something that ends in a newline (e.g., a file with a newline at the
end) and we _want_ that newline included in the data that's hashed, this
is problematic. Let's remove this trim and instead just use the `-n`
option that `echo` provides.